requirements.txt
streamlit
requests
python-dotenv
langchain-core
langchain-openai
google-generativeai